+++ src/main/java/MultiUsecaseUtil/SerializeDataAccess.java
@@ -4,7 +4,7 @@ import Entities.Group;
 import Entities.User;
 import UserRegistrationUsecase.NewUserDSGateway;
 import UserRegistrationUsecase.UserRegistrationDSRequestPackage;
-import editpendinglist.EditPendingListDsGateway;
+import edit_pending_list.EditPendingListDsGateway;
 import group_creation_use_case.GroupRegisterDSRequestModel;
 import group_creation_use_case.NewGroupDSGateway;
