+++ src/main/java/edit_pending_list/EditPendingListInputBoundary.java
@@ -1,4 +1,4 @@
-package editpendinglist;
+package edit_pending_list;
 
 public interface EditPendingListInputBoundary {
     EditPendingListResponseModel addOrRemoveUser(EditPendingListRequestModel requestModel);
