+++ src/main/java/edit_pending_list/EditPendingListRequestModel.java
@@ -1,4 +1,4 @@
-package editpendinglist;
+package edit_pending_list;
 
 public class EditPendingListRequestModel {
