+++ src/main/java/pending_list_screens/EditPendingListFailed.java
@@ -1,4 +1,4 @@
-package pendinglistscreens;
+package pending_list_screens;
 
 public class EditPendingListFailed extends RuntimeException {
     public EditPendingListFailed(String error) {
