+++ src/main/java/view_pending_list/ViewPendingListDsGateway.java
@@ -1,4 +1,4 @@
-package viewpendinglist;
+package view_pending_list;
 
 import Entities.Group;
 import Entities.User;
