+++ src/main/java/view_pending_list/ViewPendingListInputBoundary.java
@@ -1,4 +1,4 @@
-package viewpendinglist;
+package view_pending_list;
 
 public interface ViewPendingListInputBoundary {
     ViewPendingListResponseModel getUsernamesList(ViewPendingListRequestModel requestModel);
