+++ src/main/java/view_pending_list/ViewPendingListOutputBoundary.java
@@ -1,4 +1,4 @@
-package viewpendinglist;
+package view_pending_list;
 
 public interface ViewPendingListOutputBoundary {
     ViewPendingListResponseModel prepareSuccessView(ViewPendingListResponseModel usernamesList);
