+++ src/main/java/view_pending_list/ViewPendingListRequestModel.java
@@ -1,4 +1,4 @@
-package viewpendinglist;
+package view_pending_list;
 
 public class ViewPendingListRequestModel {
